+++ exe/json2yaml.hs
@@ -6,14 +6,13 @@ import qualified Data.Yaml as Y
 import Control.Monad (when)
 import qualified Data.ByteString as S
 import qualified Data.ByteString.Lazy as L
-import Control.Applicative ((<$>))
 
 main :: IO ()
 main = do
     args <- getArgs
     when (length args > 2) $ error "Usage: json2yaml [in] [out]"
     let (input:output:_) = args ++ repeat "-"
-    mval <- J.decode <$>
+    mval <- fmap J.decode $
         case input of
             "-" -> L.getContents
             _ -> L.readFile input